All the files of proof-generator and proof-generator-multithreaded deduplicated.
